fix circular dependency in ButtonTypes
