+++ src/ui/Button/types.ts
@@ -1,0 +1,10 @@
+export enum ButtonTypes {
+  PRIMARY = 'PRIMARY',
+  SECONDARY = 'SECONDARY',
+  DANGER = 'DANGER',
+  DISABLED = 'DISABLED',
+}
+export enum ButtonSizes {
+  BIG = 'BIG',
+  SMALL = 'SMALL',
+}
